Change directory and add containers
