--- src/components/Header/Header.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import './styles.css';
-
-const Header = () => <header>What the Splash!</header>;
-
-export default Header;
