+++ src/components/Header/index.js
@@ -1,3 +1,7 @@
-import Header from './Header';
+import React from 'react';
+
+import './styles.css';
+
+const Header = () => <header>What the Splash!</header>;
 
 export default Header;
